Rename shipping_common.py to shipping_helpers.py
